requirements.txt
streamlit>=1.32.0
google-genai==1.22.0
Pillow==10.1.0
python-dotenv==1.0.0
pandas==2.1.4
requests>=2.31.0
supabase>=2.8.0